Move undo-redo-history
